# apidog-import-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feedbase Backend API (APIdog Import)",
    "version": "1.0.0",
    "description": "APIdog import file generated from API_FULL_LIST.md."
  },
  "servers": [
    {
      "url": "http://localhost:4560",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api-keys/create": {
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "POST /api-keys/create",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "API key created successfully",
                  "data": {
                    "id": 1001,
                    "key": "fb_7f6312f2c6f8b8f6f9f86a7f0d7c8a8c4d9b1e0ff5f8c1234567890abcd"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "apiKeyData": {
                    "type": "object",
                    "properties": {
                      "keyName": {
                        "type": "string"
                      },
                      "scopes": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "expiresAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "apiKeyData": {
                  "keyName": "Production API Key",
                  "scopes": [
                    "read:posts",
                    "write:posts",
                    "read:analytics"
                  ],
                  "expiresAt": "2027-04-12 00:00:00"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/update/{id}": {
      "put": {
        "tags": [
          "api-keys"
        ],
        "summary": "PUT /api-keys/update/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "API key updated successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "apiKeyData": {
                    "type": "object",
                    "properties": {
                      "keyName": {
                        "type": "string"
                      },
                      "scopes": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "apiKeyData": {
                  "keyName": "Readonly Widget Key",
                  "scopes": [
                    "read:posts"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/revoke/{id}": {
      "patch": {
        "tags": [
          "api-keys"
        ],
        "summary": "PATCH /api-keys/revoke/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "API key revoked successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/api-keys/list": {
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "POST /api-keys/list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "API keys retrieved successfully",
                  "data": [
                    {
                      "id": 1001,
                      "key_name": "Production API Key",
                      "key_prefix": "fb_7f6312",
                      "scopes": "[\"read:posts\",\"write:posts\"]",
                      "last_used_at": "2026-04-12T10:00:00.000Z",
                      "expires_at": "2027-04-12T00:00:00.000Z",
                      "is_revoked": 0
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/audit-logs/list": {
      "post": {
        "tags": [
          "audit-logs"
        ],
        "summary": "POST /audit-logs/list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Audit logs retrieved successfully",
                  "data": {
                    "logs": [
                      {
                        "id": 1101,
                        "action": "POST_CREATED",
                        "entity_type": "post",
                        "entity_id": 101,
                        "metadata": "{\"title\":\"Add dark mode\"}",
                        "actor_name": "Acme Owner"
                      }
                    ],
                    "total": 97
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "paginationData": {
                    "type": "object",
                    "properties": {
                      "itemsPerPage": {
                        "type": "integer"
                      },
                      "currentPageNumber": {
                        "type": "integer"
                      },
                      "sortOrder": {
                        "type": "string"
                      },
                      "filterBy": {
                        "type": "string"
                      }
                    }
                  },
                  "filters": {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string"
                      },
                      "entityType": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "paginationData": {
                  "itemsPerPage": 20,
                  "currentPageNumber": 0,
                  "sortOrder": "desc",
                  "filterBy": ""
                },
                "filters": {
                  "action": "POST_CREATED",
                  "entityType": "post"
                }
              }
            }
          }
        }
      }
    },
    "/audit-logs/create": {
      "post": {
        "tags": [
          "audit-logs"
        ],
        "summary": "POST /audit-logs/create",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Audit log created successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "logData": {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string"
                      },
                      "entityType": {
                        "type": "string"
                      },
                      "entityId": {
                        "type": "integer"
                      },
                      "metadata": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          }
                        }
                      },
                      "ipAddress": {
                        "type": "string"
                      },
                      "userAgent": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "logData": {
                  "action": "POST_CREATED",
                  "entityType": "post",
                  "entityId": 101,
                  "metadata": {
                    "title": "Add dark mode"
                  },
                  "ipAddress": "127.0.0.1",
                  "userAgent": "Mozilla/5.0"
                }
              }
            }
          }
        }
      }
    },
...
  }
}